refactor(app): extract event handling into EventHandler
self.events = EventHandler(self) owns input routing: key handling,
paste, image drops, chip navigation, input change hints, and form
submission. Textual on_* methods on App become one-line delegates.

app.py: 1069 -> 825 lines (-244)
events.py: 303 lines (new)
