+++ src/wingman/app.py
@@ -17,7 +17,6 @@ from textual.containers import Horizontal, Vertical
 from textual.widgets import Input, Static, Tree
 
 from .checkpoints import get_checkpoint_manager
-from .command_completion import get_hint_candidates
 from .commands import Commands
 from .config import (
     APP_CREDIT,
@@ -29,9 +28,9 @@ from .config import (
     load_api_key,
 )
 from .context import AUTO_COMPACT_THRESHOLD
-from .images import cache_image_immediately, is_image_path
+from .events import EventHandler
 from .memory import load_memory
-from .sessions import load_sessions, save_session
+from .sessions import load_sessions
 from .streaming import StreamingController
 from .tools import (
     check_completed_processes,
@@ -87,6 +86,7 @@ class WingmanApp(App):
         set_app_instance(self)
         self.cmds = Commands(self)
         self.streaming = StreamingController(self)
+        self.events = EventHandler(self)
         self.scroll_sensitivity_y = 0.6
         self.client: AsyncDedalus | None = None
         self.runner: DedalusRunner | None = None
@@ -257,280 +257,36 @@ class WingmanApp(App):
         self.show_info(info)
 
     def on_descendant_focus(self, event) -> None:
-        """Set panel as active when any of its descendants receives focus."""
-        for i, panel in enumerate(self.panels):
-            if panel in event.widget.ancestors_with_self:
-                if i != self.active_panel_idx:
-                    self._set_active_panel(i)
-                break
+        self.events.on_descendant_focus(event)
 
     def on_click(self, event) -> None:
-        """Focus input when clicking anywhere in the main area."""
-        panel = self.active_panel
-        if panel:
-            # Focus the input unless clicking on an interactive element
-            from textual.widgets import Button, Input, ListView
-
-            if not isinstance(event.widget, (Button, Input, ListView, ImageChip, ToolApproval)):
-                # If there's a pending tool approval, focus that instead
-                approvals = list(panel.query("ToolApproval"))
-                if approvals:
-                    approvals[0].focus()
-                else:
-                    panel.get_input().focus()
+        self.events.on_click(event)
 
     def on_paste(self, event: events.Paste) -> None:
-        """Route paste events to the active input if not already focused there."""
-        panel = self.active_panel
-        if not panel:
-            return
-
-        input_widget = panel.get_input()
-        # If paste didn't go to the input (e.g., dropped while unfocused), route it there
-        if self.focused != input_widget and event.text:
-            # Focus the input and manually trigger paste handling
-            input_widget.focus()
-            input_widget._on_paste(event)
-            event.stop()
+        self.events.on_paste(event)
 
     @on(MultilineInput.ImageDropped)
     def on_image_dropped(self, event: MultilineInput.ImageDropped) -> None:
-        """Attach a dropped/pasted image without touching the input's typed text."""
-        panel = self.active_panel
-        if not panel:
-            return
-        image_path = is_image_path(event.path)
-        if not image_path:
-            return
-        if any(img.name == image_path.name for img in panel.pending_images):
-            return
-        cached = cache_image_immediately(image_path)
-        if not cached:
-            return
-        panel.pending_images.append(cached)
-        panel.refresh_image_chips()
-        panel.get_hint().update("[dim]↑ to select images · backspace to remove[/]")
-        self.update_status()
+        self.events.on_image_dropped(event)
 
     @on(ImageChip.Removed)
     def on_image_chip_removed(self, event: ImageChip.Removed) -> None:
-        """Remove an image when its chip is deleted."""
-        panel = self.active_panel
-        if panel and 0 <= event.index < len(panel.pending_images):
-            panel.pending_images.pop(event.index)
-            panel.refresh_image_chips()
-            self.update_status()
-            # Focus next chip or input after mount completes
-            if panel.pending_images:
-                new_idx = min(event.index, len(panel.pending_images) - 1)
-
-                def focus_chip():
-                    chips = list(panel.get_chips_container().query(ImageChip))
-                    if chips and new_idx < len(chips):
-                        chips[new_idx].focus()
-
-                self.call_after_refresh(focus_chip)
-            else:
-                panel.get_hint().update("")
-                panel.get_input().focus()
+        self.events.on_image_chip_removed(event)
 
     @on(ImageChip.Navigate)
     def on_image_chip_navigate(self, event: ImageChip.Navigate) -> None:
-        """Handle chip navigation."""
-        panel = self.active_panel
-        if not panel:
-            return
-        chips = list(panel.get_chips_container().query(ImageChip))
-        if event.direction == "down":
-            panel.get_input().focus()
-        elif event.direction == "left" and event.index > 0:
-            chips[event.index - 1].focus()
-        elif event.direction == "right" and event.index < len(chips) - 1:
-            chips[event.index + 1].focus()
+        self.events.on_image_chip_navigate(event)
 
     def on_key(self, event) -> None:
-        """Handle escape and arrow navigation for image chips."""
-        # Handle escape only when no modal is open
-        if event.key == "escape" and len(self.screen_stack) == 1:
-            panel = self.active_panel
-            if panel and panel._generating:
-                panel._cancel_requested = True
-                panel._generating = False  # Clear immediately
-                self.update_status()
-                # Remove thinking spinners
-                for thinking in panel.query("Thinking"):
-                    with contextlib.suppress(Exception):
-                        thinking.remove()
-                # Remove pending tool approvals
-                for approval in panel.query("ToolApproval"):
-                    with contextlib.suppress(Exception):
-                        approval.remove()
-                self.notify("Generation cancelled", severity="warning", timeout=2)
-                event.stop()
-                event.prevent_default()
-                return
-            elif panel:
-                try:
-                    input_widget = panel.query_one(f"#{panel.panel_id}-prompt", Input)
-                    input_widget.value = ""
-                    if hasattr(input_widget, "_pasted_content"):
-                        input_widget._pasted_content = None
-                        input_widget._paste_placeholder = None
-                except Exception:
-                    pass
-                event.stop()
-                event.prevent_default()
-                return
-
-        panel = self.active_panel
-        if not panel:
-            return
-
-        focused = self.focused
-
-        # Up from input -> last chip
-        if event.key == "up" and panel.pending_images:
-            if focused and isinstance(focused, Input) and "panel-prompt" in focused.classes:
-                chips = list(panel.get_chips_container().query(ImageChip))
-                if chips:
-                    event.prevent_default()
-                    chips[-1].focus()
-
-        # Navigation when chip is focused
-        elif isinstance(focused, ImageChip):
-            chips = list(panel.get_chips_container().query(ImageChip))
-            try:
-                idx = chips.index(focused)
-            except ValueError:
-                return
-
-            if event.key == "down":
-                event.prevent_default()
-                panel.get_input().focus()
-            elif event.key == "left" and idx > 0:
-                event.prevent_default()
-                chips[idx - 1].focus()
-            elif event.key == "right" and idx < len(chips) - 1:
-                event.prevent_default()
-                chips[idx + 1].focus()
+        self.events.on_key(event)
 
     @on(Input.Changed, ".panel-prompt")
     def on_input_changed(self, event: Input.Changed) -> None:
-        """Show command hints when typing / and auto-detect image paths."""
-        panel = None
-        for ancestor in event.input.ancestors_with_self:
-            if isinstance(ancestor, ChatPanel):
-                panel = ancestor
-                break
-        if not panel:
-            return
-        hint = panel.get_hint()
-        text = event.value
-
-        # Auto-detect image paths (drag-and-drop)
-        # Check for image extensions in various formats (plain, URL-encoded, backslash-escaped)
-        text_lower = text.strip().strip("'\"").lower() if text else ""
-        has_image_ext = any(
-            text_lower.endswith(ext) or text_lower.endswith(ext.replace(".", "%2e"))
-            for ext in (".png", ".jpg", ".jpeg", ".gif", ".webp", ".bmp")
-        )
-        if text and has_image_ext:
-            image_path = is_image_path(text)
-            if image_path:
-                # Prevent duplicate adds from rapid-fire input events
-                if any(img.name == image_path.name for img in panel.pending_images):
-                    event.input.clear()
-                    return
-                cached = cache_image_immediately(image_path)
-                if cached:
-                    event.input.clear()
-                    panel.pending_images.append(cached)
-                    panel.refresh_image_chips()
-                    panel.get_hint().update("[dim]↑ to select images · backspace to remove[/]")
-                    self.update_status()
-                    return
-
-        if text.startswith("/"):
-            # Don't overwrite hint if actively cycling through completions for this exact input
-            cycle = getattr(event.input, "_completion_cycle", None)
-            if cycle and cycle.is_active_for(text, event.input.cursor_position):
-                return
-            matches = get_hint_candidates(text, event.input.cursor_position)
-            formatted = "  ".join(f"[#7aa2f7]{cmd}[/]" for cmd in matches)
-            hint.update(formatted if formatted else "")
-        elif panel.pending_images:
-            hint.update("[dim]↑ to select images · backspace to remove[/]")
-        else:
-            hint.update("")
+        self.events.on_input_changed(event)
 
     @on(Input.Submitted, ".panel-prompt")
     def on_submit(self, event: Input.Submitted) -> None:
-        panel = None
-        for p in self.panels:
-            if p.panel_id in event.input.id:
-                panel = p
-                break
-        if not panel:
-            return
-
-        # Block input while generating
-        if panel._generating:
-            self.notify("Wait for response to complete", severity="warning", timeout=2)
-            return
-
-        # Activate this panel if it's not active
-        if panel != self.active_panel:
-            self._set_active_panel(self.panels.index(panel))
-
-        text = (
-            event.input.get_submit_value().strip() if hasattr(event.input, "get_submit_value") else event.value.strip()
-        )
-
-        if not text and not panel.pending_images:
-            return
-
-        event.input.clear()
-        panel.get_hint().update("")
-
-        if text.startswith("/"):
-            self.handle_command(text)
-            return
-
-        # Remove welcome message if present
-        try:
-            for child in panel.get_chat_container().children:
-                if "panel-welcome" in child.classes:
-                    child.remove()
-                    break
-        except Exception:
-            pass
-
-        # Create new session if none exists
-        if not panel.session_id:
-            panel.session_id = f"chat-{int(time.time() * 1000)}"
-            save_session(panel.session_id, [])
-            self.refresh_sessions()
-            self.update_status()
-
-        # Handle images in message
-        images_to_send = panel.pending_images.copy()
-        panel.pending_images = []
-        panel.refresh_image_chips()  # Clear chips display
-
-        if images_to_send:
-            panel.add_image_message("user", text, images_to_send)
-        else:
-            panel.add_message("user", text)
-
-        # Save session immediately after user message
-        save_session(panel.session_id, panel.messages)
-
-        chat = panel.get_chat_container()
-        thinking = Thinking(id="thinking")
-        chat.mount(thinking)
-        panel.get_scroll_container().scroll_end(animate=False)
-
-        self.streaming.send_message(panel, text, thinking, images_to_send)
+        self.events.on_submit(event)
 
     def show_diff_approval(self) -> None:
         """Show diff modal for pending edit approval. Called from tool thread."""
@@ -648,7 +404,7 @@ class WingmanApp(App):
         sidebar = self.query_one("#sidebar")
         sidebar.display = not sidebar.display
 
-    def _set_active_panel(self, idx: int) -> None:
+    def set_active_panel(self, idx: int) -> None:
         """Set the active panel by index."""
         if idx < 0 or idx >= len(self.panels):
             return
@@ -673,7 +429,7 @@ class WingmanApp(App):
         # Refresh welcome art on existing panels after layout recalculates
         self.call_after_refresh(self._refresh_welcome_art)
         # Activate the new panel
-        self._set_active_panel(len(self.panels) - 1)
+        self.set_active_panel(len(self.panels) - 1)
         self.update_status()
 
     def _refresh_welcome_art(self) -> None:
@@ -733,30 +489,30 @@ class WingmanApp(App):
         if len(self.panels) <= 1:
             return
         new_idx = (self.active_panel_idx - 1) % len(self.panels)
-        self._set_active_panel(new_idx)
+        self.set_active_panel(new_idx)
 
     def action_next_panel(self) -> None:
         """Switch to next panel."""
         if len(self.panels) <= 1:
             return
         new_idx = (self.active_panel_idx + 1) % len(self.panels)
-        self._set_active_panel(new_idx)
+        self.set_active_panel(new_idx)
 
     def action_goto_panel_1(self) -> None:
         if len(self.panels) >= 1:
-            self._set_active_panel(0)
+            self.set_active_panel(0)
 
     def action_goto_panel_2(self) -> None:
         if len(self.panels) >= 2:
-            self._set_active_panel(1)
+            self.set_active_panel(1)
 
     def action_goto_panel_3(self) -> None:
         if len(self.panels) >= 3:
-            self._set_active_panel(2)
+            self.set_active_panel(2)
 
     def action_goto_panel_4(self) -> None:
         if len(self.panels) >= 4:
-            self._set_active_panel(3)
+            self.set_active_panel(3)
 
     def _mount_command_status(self, command: str, widget_id: str, panel_id: str | None = None) -> None:
         """Mount command status widget in the specified panel, before Thinking spinner."""
